Starting work on refactoring entire project
A lot of changes are being made which pretty much involves the whole
package being refactored. Keeping old methods and tests intact so that
tests could run and the package is still usable. However, the driver
has been changed and the README must also be rewritten.

This commit contains a new type, Runner, and tests. Can create simple
files. Again, file creation is simply limited to a touch command for now.